Extract buffered reader
