+++ paddle/fluid/operators/reader/create_double_buffer_reader_op.cc
@@ -12,83 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <thread>  // NOLINT
-
-#include "ThreadPool.h"
-#include "paddle/fluid/operators/reader/blocking_queue.h"
+#include "paddle/fluid/operators/reader/buffered_reader.h"
 #include "paddle/fluid/operators/reader/reader_op_registry.h"
 
 namespace paddle {
 namespace operators {
 namespace reader {
-class BufferedReader : public framework::DecoratedReader {
-  using TensorVec = std::vector<framework::LoDTensor>;
-  using VecFuture = std::future<TensorVec>;
-
- public:
-  BufferedReader(const std::shared_ptr<framework::ReaderBase>& reader,
-                 const platform::Place& place, size_t buffer_size)
-      : framework::DecoratedReader(reader),
-        thread_pool_(1),
-        place_(place),
-        buffer_size_(buffer_size) {
-    AppendFutureToBatchSize();
-  }
-
-  ~BufferedReader() override {
-    reader_->Shutdown();
-    buffer_.clear();
-  }
-
- private:
-  void AppendFutureToBatchSize() {
-    while (buffer_.size() < buffer_size_) {
-      AppendFuture();
-    }
-  }
-
-  void AppendFuture() {
-    buffer_.emplace_back(thread_pool_.enqueue([this] {
-      TensorVec cpu_buffer;
-      reader_->ReadNext(&cpu_buffer);
-      if (platform::is_gpu_place(place_)) {
-        TensorVec gpu_buffer;
-
-        for (size_t i = 0; i < cpu_buffer.size(); ++i) {
-          gpu_buffer.emplace_back();
-          framework::TensorCopySync(cpu_buffer[i], place_, &gpu_buffer.back());
-        }
-
-        cpu_buffer = gpu_buffer;
-      }
-      return cpu_buffer;
-    }));
-  }
-
- protected:
-  void ShutdownImpl() override {
-    reader_->Shutdown();
-    buffer_.clear();
-  }
-  void StartImpl() override {
-    reader_->Start();
-    AppendFutureToBatchSize();
-  }
-  void ReadNextImpl(std::vector<framework::LoDTensor>* out) override {
-    std::cerr << "Read" << std::endl;
-    PADDLE_ENFORCE_EQ(buffer_.size(), buffer_size_);
-    *out = buffer_.front().get();
-    buffer_.pop_front();
-    AppendFuture();
-  }
-
- private:
-  ThreadPool thread_pool_;
-  platform::Place place_;
-  const size_t buffer_size_;
-  std::list<VecFuture> buffer_;
-};
-
 class CreateDoubleBufferReaderOp : public framework::OperatorBase {
  public:
   using framework::OperatorBase::OperatorBase;
